Perform login step to currently fixed url http://localhost:8002 at last position.
This only performs the login step.  There is absolutely nothing to actually connect to the target simulator which is the next, much harder step.
This currently requires the patch at https://gist.github.com/justincc/74305bb0b1a84f25c5e5 to be applied to OpenSimulator in order to work.
Sometimes on Firefox the CORs doesn't work properly and proper requests are only generated for the first attempt.  Temporary workaround is to restart the browser.
Only tested on Firefox 33.0